Fixed tests to accomadate new file StrongerBrick
Had to put StrongerBrick in a new file because Brick would not export both classes, would love to learn to make that cleaner if possible. But game properly controls the creation of bricks now.
