+++ lib/game.js
@@ -1,4 +1,5 @@
 const Brick = require('./Brick.js');
+const StrongerBrick = require('./StrongerBrick.js');
 
 class Game {
   constructor(ball, paddle) {
